refactor: remove duplicate metrics aggregator service from rqueue-web
Clean up duplicate RqueueJobMetricsAggregatorService from rqueue-web/src/main/java after moving the interface to rqueue-core/metrics as RqueueMetricsAggregatorService.

Move the test RqueueTaskMetricsAggregatorServiceTest from rqueue-web to rqueue-core/src/test/java/com/github/sonus21/rqueue/metrics/, updating package declarations and class references to use the renamed RqueueMetricsAggregatorService.

This ensures a single source of truth for the metrics aggregator interface in the core module.
